chore(util): centralize date/time serializer and helpers (#7)
