+++ src/main/kotlin/org/darren/stock/domain/actors/events/StockPotEvent.kt
@@ -3,7 +3,7 @@ package org.darren.stock.domain.actors.events
 import kotlinx.serialization.Serializable
 import org.darren.stock.domain.DateTimeProvider
 import org.darren.stock.domain.StockState
-import org.darren.stock.ktor.DateSerializer
+import org.darren.stock.util.DateSerializer
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import java.time.LocalDateTime
