+++ src/main/kotlin/org/darren/stock/ktor/Sale.kt
@@ -41,7 +41,7 @@ object Sale {
     private data class SaleRequestDTO(
         val requestId: String,
         val quantity: Double,
-        @Serializable(with = DateSerializer::class) val soldAt: LocalDateTime,
+        @Serializable(with = org.darren.stock.util.DateSerializer::class) val soldAt: LocalDateTime,
     )
 
     @Serializable
